Validate `pytest-mypy-plugins` input file schema
Create a `schema.json` to:
* Validate the input provided by the users
* Offer in-editor validation and auto-completion
* Easily keep the documentation of it up-to-date

Use said schema to meta-test all test files for conformance.

Additionally:
* Fix `mypy_config` type to `str | None`
* Update `jinja2.defaults.VARIABLE_START_STRING` to the
  more-correct `_rendering_env.variable_start_string`.
* Update `.gitignore`

This fixes the real issue behind https://github.com/typeddjango/pytest-mypy-plugins/pull/124:
The problem was not that `mypy_config` *MUST HAVE* `{{` when `parametrized` was set;
It was passing a `list` (of `dict`s) - which that was not templatable.

Signed-off-by: Stavros Ntentos <133706+stdedos@users.noreply.github.com>